[init] Fix signness warning

+++ user/init/src/main.c
@@ -43,11 +43,11 @@ void _start(int argc, const char** argv, int envc, const char** envv) {
     _libc_init_environ(envv, envc);
     _libc_init_streams();
     printf("Args dump:\n");
-    for(size_t i = 0; i < argc; ++i) {
+    for(size_t i = 0; i < (size_t)argc; ++i) {
         printf("%zu> ",i+1); printf("%p",argv[i]); printf(" %s\n",argv[i]);
     }
     printf("Env dump:\n");
-    for(size_t i = 0; i < envc; ++i) {
+    for(size_t i = 0; i < (size_t)envc; ++i) {
         printf("%zu> ",i+1); printf("%p",envv[i]); printf(" %s\n",envv[i]);
     }
     int code = main();
